create prj

+++ prjWebBanHang/database/migrations/2024_11_06_124644_tbl_order_details.php
@@ -15,11 +15,12 @@ return new class extends Migration
     {
         Schema::create('tbl_order_details', function (Blueprint $table) {
             $table->bigIncrements('order_details_id');
-            $table->integer('order_id');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('order_id');
+            $table->unsignedInteger('product_id');
             $table->string('product_name');
             $table->float('product_price');
             $table->integer('product_sales_quantity');
+            
             $table->timestamps();
         });
     }
